Bug 7534: Add new column "pickup_location" to branches
This patch adds a new column "pickup_location" for each library. This value
defines whether a library can be used as a pickup location or not.

For extended configurability, this flag could perhaps be considered as a part
of issuing rules, but as of now, this patch only offers a simple flag into
each library. This method also allows easy way for REST API to determine valid
pickup locations by querying the /libraries endpoint.

To test:
1. perl installer/data/mysql/updatedatabase.pl
2. Observe
DEV atomic update: bug_7534.perl
Upgrade to XXX done (Bug 7534 - Let libraries have configuration for pickup locations)

https://bugs.koha-community.org/show_bug.cgi?id=7354

Signed-off-by: Koha Team AMU <axelle.clarisse@univ-amu.fr>