Bug 18768 - Add support for authenticating against external SIP servers ala LDAP and CAS
This feature allows Koha use to use other ILSs as an authenticate target in a manner someone similar to how we authenticate LDAP. It allows Koha to be configured with a list of external SIP2 servers to authenticate against. Koha will find the correct server to use based on the order and cardnumber prefix chosen, send a patron information request and handle the patron login based on the response.

Test Plan:
1) Apply this patch set
2) Set up another Koha server as a SIP2 target
3) Follow the setup procedure detailed in the POD in Auth_with_sip.pm
4) Log in with a username/password that exists on the external server but not in this one