Bug 18816 - Make CataloguingLog work in production by preventing circulation from spamming the log
The system preference CataloguingLog is not recommended for use in
production. This is do to the fact that every checkin and checkout
generates one or more log entires. This seems to be not only bad
behavior, but unnecessary and outside the needs of CataloguingLog as we
have CirculationLog.

Test Plan:
1) Enable CataloguingLog
2) Note the checkins and checkouts cause log entries
3) Apply this patch
4) Note that checkins and checkouts no longer cause log entries,
   but editing the item will continue to cause log entries.