Bug XXX - Allow LDAP mapped fields to be skipped for updates
Sometimes LDAP replication brings over 'default' values for a field that
the library will change afterward. If replicate and update are both enabled
those changed fields will be overwritten each time the user logs in.

It would be nice to be able to specify that a field should be set from
the LDAP source for replication, but not for updates.