Bug 20813: Revamp user permissions system
Koha's current user permissions system is a bit convoluted and limited
due to it's legacy where the permissions system was a simple set of
flags. A second layer of sub-permissions were later added complication
and obfuscation to the system.

We should revamp the permissions system to be a more modern system and
open the path to having unlimited nesting of sub-permissions instead of
shoe-horning sub-permissions into the limited system we have.

Test Plan:
1) Apply this patch set
2) Run updatedatabase.pl
3) Restart all the things!
4) Koha should still enforce user permissions as before