Bug 17378: Add ability to specify maximum number of items per record
Some libraries want to be able to specify a hard limit on the number of items a record can have. It would be nice to have this set per record type, but since most libraries use item level itemtypes, that would make it confusing and non-intuitive. The simplest and most straight forward approach would just be to have a system preference. Koha should then just report that it failed to create an item if the librarian has hit the upper limit specified in the system preference.

Test Plan:
1) Apply this patch
2) Run updatedatabase.pl
3) Set the new syspref RecordItemLimit to 3
4) Find a record with 4 or more records
5) Note in the catalog toolbar you cannot choose the add new item option
6) Note in the items editor you can edit items but not add new ones