Hiding the 'faux' analytics
Hiding the 'analytics' that appear when a site places s in the leader.  This was put it to support consortial libraries that share records, but it is not correct cataloging.

Removed 'fix' for hiding analytics

This change appears to have caused all marc to disappear on staff side when i did it.  oops.

Hiding faux analytics

Hide analytics link based on leader pos 7=s

Remove faux analytics

where leader pos7=s