Bug 21951: Patron import fails if matching on cardnumber and userid also matches
If a patron CSV file has both cardnumber and userid, and cardnumber is used as a matchpoint, the patron import will fail with an error like
--
Userid kyle.hall is already used by another patron.
--
even though that userid is used by the very same patron!

Matching on userid does not generate this error.

Test Plan:
1) Generate a patron CSV file where the userid and cardnumber columns match an existing patron
2) Attempt to overlay existing patron using this CSV file via the patron import tool
3) Note the "already used" error
4) Apply this patch
5) No error should be generated
6) Modify the cardnumber in the CSV file
7) Attempt to import and overlay again
8) Note you get the userid "already used" error