Bug 22173: Search Groups broken by change in expected structure
The original logic of Search Groups using the heirarchical library groups was to enable search groups on a root group and have each sub group be a search group listed in the pulldown:

RootGroup:
   SearchGroup1:
        LibraryA
        LibraryB
   SearchGroup2:
        LibraryC
        LibraryD
   SearchGroup3:
        LibraryE
        LibraryF

At some point, the code was changed to expect the each search group to be the root group:

RootSearchGroup1:
    LibraryA
    LibraryB

RootSearchGroup2:
    LibraryC
    LibraryD

RootSearchGroup3:
    LibraryE
    LibraryF

It's likely that both styles are being used in production now with different versions. If a library upgrades, their search groups will be broken. If we revert the change, currently up to date libraries will have search groups broken.

Instead, we should accommodate both configurations.

So, if a root group has libraries as direct children, it should be displayed as a search group. If it does not, we should display the child search groups instead.

Test Plan:
1) Set up a root group with multiple subgroups each with libraries,
   enable staff and opac search groups for that root group
2) Note the root group is displayed in all the search pulldowns instead
   of the subgroups
3) Apply this patch
4) Restart all the things!
5) Note the search pulldowns now contain the subgroups instead
6) Create another root group, add only libraries to it.
   Set it as a search group for staff and opac.
7) Note the search pulldowns now contain this group as well!