Bug 22200 - Forgiving a fine (FFOR) does not update the accountline date
Each time a fine is incremented, the date column is updated. When that fine is forgiven ( often via the "Forgive fines" option on the checkin page ), it does not update the column. It makes sense that if updating a fine sets the date column, forgiving it should do so as well.

Test Plan:
1) Find a checkout that is overdue with fines, where the date is not today
2) Note the date in the accountlines table
3) Check it in, note the date is not updated
4) Apply this patch
5) Repeat steps 1-2
6) Note the date is updated!